move the main project source into `Enhancer` project directory
