--- Enhancer.csproj
@@ -1,60 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net472</TargetFramework>
-    <AssemblyName>Enhancer</AssemblyName>
-    <Description>Tweaks plugin for Lethal Company</Description>
-    <Version>0.1.4</Version>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <LangVersion>latest</LangVersion>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <OutputPath>./bin/Flowerful</OutputPath>
-    <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-  </PropertyGroup>
-
-  <!-- make all references non-private, so they won't be copied to the output folder -->
-  <Target Name="ClearReferenceCopyLocalPaths" AfterTargets="ResolveAssemblyReferences">
-    <ItemGroup>
-      <ReferenceCopyLocalPaths Remove="@(ReferenceCopyLocalPaths)" />
-    </ItemGroup>
-  </Target>
-
-  <ItemGroup>
-    <PackageReference Include="BepInEx.Analyzers" Version="1.*" PrivateAssets="all" />
-    <PackageReference Include="BepInEx.Core" Version="5.*" />
-    <PackageReference Include="BepInEx.PluginInfoProps" Version="1.*" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies.net472" Version="1.0.3">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-    </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="Assembly-CSharp">
-      <HintPath>../lib/Assembly-CSharp.dll</HintPath>
-    </Reference>
-    <Reference Include="Unity Netcode">
-      <HintPath>../lib/Unity.Netcode.Runtime.dll</HintPath>
-    </Reference>
-    <Reference Include="Netstandard">
-      <HintPath>../lib/netstandard.dll</HintPath>
-    </Reference>
-    <Reference Include="UnityEngine UIModule">
-      <HintPath>../lib/UnityEngine.UIModule.dll</HintPath>
-    </Reference>
-    <Reference Include="UnityEngine Core">
-      <HintPath>../lib/UnityEngine.CoreModule.dll</HintPath>
-    </Reference>
-    <Reference Include="UnityEngine">
-      <HintPath>../lib/UnityEngine.dll</HintPath>
-    </Reference>
-    <Reference Include="UnityEngine.TMP">
-      <HintPath>../lib/Unity.TextMeshPro.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-</Project>
